pyproject.toml
[tool.poetry]
name = "telegram_exception_alerts"
version = "0.2.0"
description = ""
authors = ["licht1stein <mb@blaster.ai>"]
homepage = "https://github.com/licht1stein/telegram-exception-alerts"
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3.0"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
